Refactoring balance calculator switch statement and add contributions handler class.
