+++ src/test/java/uk/gov/homeoffice/digital/sas/balancecalculator/BalanceCalculatorTest.java
@@ -1,5 +1,26 @@
 package uk.gov.homeoffice.digital.sas.balancecalculator;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+import static uk.gov.homeoffice.digital.sas.balancecalculator.BalanceCalculator.ACCRUALS_NOT_FOUND;
+import static uk.gov.homeoffice.digital.sas.balancecalculator.BalanceCalculator.AGREEMENT_NOT_FOUND;
+import static uk.gov.homeoffice.digital.sas.balancecalculator.handlers.ContributionsHandler.MISSING_ACCRUAL;
+import static uk.gov.homeoffice.digital.sas.balancecalculator.testutils.CommonUtils.createAccrual;
+import static uk.gov.homeoffice.digital.sas.balancecalculator.testutils.CommonUtils.loadAccrualsFromFile;
+import static uk.gov.homeoffice.digital.sas.balancecalculator.testutils.CommonUtils.loadObjectFromFile;
+
+import java.io.IOException;
+import java.math.BigDecimal;
+import java.text.MessageFormat;
+import java.time.LocalDate;
+import java.time.ZonedDateTime;
+import java.util.List;
+import java.util.Map;
+import java.util.SortedMap;
+import java.util.UUID;
+import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -11,43 +32,15 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.boot.test.system.CapturedOutput;
 import org.springframework.boot.test.system.OutputCaptureExtension;
 import uk.gov.homeoffice.digital.sas.balancecalculator.client.RestClient;
+import uk.gov.homeoffice.digital.sas.balancecalculator.handlers.ContributionsHandler;
 import uk.gov.homeoffice.digital.sas.balancecalculator.models.accrual.Accrual;
 import uk.gov.homeoffice.digital.sas.balancecalculator.models.accrual.Agreement;
-import uk.gov.homeoffice.digital.sas.balancecalculator.models.accrual.Contributions;
 import uk.gov.homeoffice.digital.sas.balancecalculator.models.accrual.enums.AccrualType;
 import uk.gov.homeoffice.digital.sas.balancecalculator.models.timecard.TimeEntry;
 import uk.gov.homeoffice.digital.sas.balancecalculator.module.AccrualModule;
 import uk.gov.homeoffice.digital.sas.balancecalculator.module.AnnualTargetHoursAccrualModule;
 import uk.gov.homeoffice.digital.sas.balancecalculator.testutils.CommonUtils;
 import uk.gov.homeoffice.digital.sas.kafka.message.KafkaAction;
-
-import java.io.IOException;
-import java.math.BigDecimal;
-import java.text.MessageFormat;
-import java.time.LocalDate;
-import java.time.ZonedDateTime;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.SortedMap;
-import java.util.TreeMap;
-import java.util.UUID;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static uk.gov.homeoffice.digital.sas.balancecalculator.BalanceCalculator.ACCRUALS_MAP_EMPTY;
-import static uk.gov.homeoffice.digital.sas.balancecalculator.BalanceCalculator.ACCRUALS_NOT_FOUND;
-import static uk.gov.homeoffice.digital.sas.balancecalculator.BalanceCalculator.AGREEMENT_NOT_FOUND;
-import static uk.gov.homeoffice.digital.sas.balancecalculator.BalanceCalculator.MISSING_ACCRUAL;
-import static uk.gov.homeoffice.digital.sas.balancecalculator.testutils.CommonUtils.createAccrual;
-import static uk.gov.homeoffice.digital.sas.balancecalculator.testutils.CommonUtils.loadAccrualsFromFile;
-import static uk.gov.homeoffice.digital.sas.balancecalculator.testutils.CommonUtils.loadObjectFromFile;
 
 @ExtendWith({MockitoExtension.class, OutputCaptureExtension.class})
 class BalanceCalculatorTest {
@@ -68,6 +61,8 @@ class BalanceCalculatorTest {
   private RestClient restClient;
 
   private BalanceCalculator balanceCalculator;
+
+  private ContributionsHandler contributionsHandler;
 
   private static Stream<Arguments> annualTargetHoursTestData() {
     return Stream.of(
@@ -104,7 +99,8 @@ class BalanceCalculatorTest {
 
   @BeforeEach
   void setup() {
-    balanceCalculator = new BalanceCalculator(restClient, accrualModules);
+    contributionsHandler = new ContributionsHandler();
+    balanceCalculator = new BalanceCalculator(restClient, accrualModules, contributionsHandler);
   }
 
   @Test
@@ -269,98 +265,5 @@ class BalanceCalculatorTest {
         .isEqualTo(BigDecimal.valueOf(8040));
   }
 
-  @Test
-  void cascadeCumulativeTotal_priorDateWithinSameAgreement_usePriorCumulativeTotalAsBasis()
-      throws IOException {
-    List<Accrual> accruals = loadAccrualsFromFile("data/accruals_annualTargetHours.json");
 
-    SortedMap<LocalDate, Accrual> map = accruals.stream()
-        .collect(Collectors.toMap(
-            Accrual::getAccrualDate,
-            Function.identity(),
-            (k1, k2) -> k2,
-            TreeMap::new)
-        );
-
-    LocalDate agreementStartDate = LocalDate.of(2023, 4, 1);
-    LocalDate referenceDate = LocalDate.of(2023, 4, 18);
-
-    balanceCalculator.cascadeCumulativeTotal(map, agreementStartDate);
-
-    assertThat(map).hasSize(5);
-    assertThat(map.get(referenceDate)
-        .getCumulativeTotal()).usingComparator(BigDecimal::compareTo)
-        .isEqualTo(BigDecimal.valueOf(6480));
-
-    assertThat(map.get(referenceDate.plusDays(1))
-        .getCumulativeTotal()).usingComparator(BigDecimal::compareTo)
-        .isEqualTo(BigDecimal.valueOf(7080));
-
-    assertThat(map.get(referenceDate.plusDays(2))
-        .getCumulativeTotal()).usingComparator(BigDecimal::compareTo)
-        .isEqualTo(BigDecimal.valueOf(7320));
-
-    assertThat(map.get(referenceDate.plusDays(3))
-        .getCumulativeTotal()).usingComparator(BigDecimal::compareTo)
-        .isEqualTo(BigDecimal.valueOf(8040));
-  }
-
-  @Test
-  void cascadeCumulativeTotal_emptyMapOfAccruals_throwException() {
-    SortedMap<LocalDate, Accrual> map = new TreeMap<>();
-
-    LocalDate agreementStartDate = LocalDate.of(2023, 4, 1);
-
-    assertThatThrownBy(() ->
-        balanceCalculator.cascadeCumulativeTotal(map, agreementStartDate))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessageContaining(ACCRUALS_MAP_EMPTY);
-  }
-
-  @Test
-  void updateAccrualContribution_hasNoContributions_returnUpdatedAccrual() {
-
-    Contributions contributions = Contributions.builder()
-        .timeEntries(new HashMap<>())
-        .total(BigDecimal.ZERO)
-        .build();
-    Accrual accrual = Accrual.builder()
-        .accrualDate(ACCRUAL_DATE)
-        .contributions(contributions)
-        .build();
-
-    balanceCalculator.updateAccrualContribution(TIME_ENTRY_ID, SHIFT_DURATION, accrual);
-
-    assertThat(accrual.getContributions().getTotal()).usingComparator(BigDecimal::compareTo)
-        .isEqualTo(BigDecimal.valueOf(120));
-    assertThat(accrual.getContributions().getTimeEntries()).hasSize(1);
-    assertThat(accrual.getContributions().getTimeEntries()).containsEntry(
-        UUID.fromString(TIME_ENTRY_ID), SHIFT_DURATION);
-  }
-
-  @Test
-  void updateAccrualContribution_hasExistingContribution_returnUpdatedAccrual() {
-    UUID existingTimeEntryId = UUID.fromString("e7d85e42-f0fb-4e2a-8211-874e27d1e888");
-
-    Contributions contributions = Contributions.builder()
-        .timeEntries(new HashMap<>(Map.of(existingTimeEntryId, BigDecimal.TEN)))
-        .total(BigDecimal.TEN)
-        .build();
-    Accrual accrual = Accrual.builder()
-        .accrualDate(ACCRUAL_DATE)
-        .contributions(contributions)
-        .build();
-
-    balanceCalculator.updateAccrualContribution(TIME_ENTRY_ID, SHIFT_DURATION, accrual);
-
-    assertThat(accrual.getContributions().getTotal()).usingComparator(BigDecimal::compareTo)
-        .isEqualTo(BigDecimal.valueOf(130));
-
-    assertThat(accrual.getContributions().getTimeEntries()).hasSize(2);
-
-    assertThat(accrual.getContributions().getTimeEntries()).containsEntry(
-        existingTimeEntryId, BigDecimal.TEN);
-    assertThat(accrual.getContributions().getTimeEntries()).containsEntry(
-        UUID.fromString(TIME_ENTRY_ID), SHIFT_DURATION);
-  }
 }
